requirements.txt

nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2

nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
setuptools==77.0.3
six==1.17.0
sympy==1.13.1
tifffile==2025.3.13
torch==2.6.0
torchvision==0.21.0

typing_extensions==4.12.2
